# docs/openapi-c-v1.yaml
openapi: 3.0.3
info:
  title: Baoyunlian C API
  version: 1.0.0
  description: Frozen C-side API contract for frontend-backend integration.
servers:
  - url: http://127.0.0.1:4000
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/send-code:
    post:
      tags: [Auth]
      summary: Send SMS code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mobile: { type: string, pattern: '^1[3-9]\\d{9}$' }
              required: [mobile]
      responses:
        '200':
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
                  dev_code: { type: string }
                required: [ok, message]
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/auth/verify-basic:
    post:
      tags: [Auth]
      summary: Verify basic identity and login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 2, maxLength: 20 }
                mobile: { type: string, pattern: '^1[3-9]\\d{9}$' }
                code: { type: string, pattern: '^\\d{6}$' }
              required: [name, mobile, code]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
                required: [token, user]
        '400': { $ref: '#/components/responses/BadRequest' }
...
  /api/learning/courses/{id}:
    get:
      tags: [Learning]
      summary: Learning course detail
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  course: { $ref: '#/components/schemas/LearningCourse' }
                required: [course]
        '404': { $ref: '#/components/responses/NotFound' }
  /api/learning/courses/{id}/complete:
    post:
      tags: [Learning]
      summary: Complete course and earn points
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  duplicated: { type: boolean }
                  reward: { type: integer }
                  balance: { type: integer }
                  message: { type: string }
                required: [ok, duplicated, reward, balance]
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/insurance/policies:
    get:
      tags: [Insurance]
      summary: Insurance policies
      security: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items: { $ref: '#/components/schemas/InsurancePolicy' }
                required: [policies]

    post:
      tags: [Insurance]
      summary: Create insurance policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company: { type: string }
                name: { type: string }
                applicant: { type: string }
                insured: { type: string }
                date: { type: string }
                paymentPeriod: { type: string }
                coveragePeriod: { type: string }
                amount: { type: integer }
                firstPremium: { type: integer }
                type: { type: string }
              required: [company, name, applicant, insured, date, paymentPeriod, coveragePeriod, amount, firstPremium]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  policy: { $ref: '#/components/schemas/InsurancePolicy' }
                required: [ok, policy]
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/insurance/policies/{id}:
    get:
      tags: [Insurance]
      summary: Insurance policy detail
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  policy: { $ref: '#/components/schemas/InsurancePolicy' }
                required: [policy]
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
...
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
  schemas:
    ErrorBody:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
      required: [code, message]
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        mobile: { type: string }
        is_verified_basic: { type: boolean }
        verified_at: { type: string, nullable: true }
      required: [id, name, mobile, is_verified_basic]
...
    LearningCourse:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        desc: { type: string }
        type:
          type: string
          enum: [video, comic, article]
        typeLabel: { type: string }
        progress: { type: integer }
        timeLeft: { type: string }
        image: { type: string }
        action: { type: string }
        color: { type: string }
        btnColor: { type: string }
        points: { type: integer }
        category: { type: string }
        content: { type: string }
      required:
        - id
        - title
        - desc
        - type
        - typeLabel
        - progress
        - timeLeft
        - image
        - action
        - color
        - btnColor
        - points
        - category
        - content
    InsurancePolicy:
      type: object
      properties:
        id: { type: integer }
        company: { type: string }
        name: { type: string }
        type: { type: string }
        icon:
          type: string
          enum: [stethoscope, heart-pulse, shield]
        amount: { type: integer }
        nextPayment: { type: string }
        status: { type: string }
        applicant: { type: string }
        insured: { type: string }
        periodStart: { type: string }
        periodEnd: { type: string }
        annualPremium: { type: integer }
        paymentPeriod: { type: string }
        coveragePeriod: { type: string }
        responsibilities:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              desc: { type: string }
              limit: { type: integer }
            required: [name, desc, limit]
        paymentHistory:
          type: array
          items:
            type: object
            properties:
              date: { type: string }
              amount: { type: integer }
              note: { type: string }
              status: { type: string }
            required: [date, amount, note, status]
        policyNo: { type: string }
      required:
        - id
        - company
        - name
        - type
        - icon
        - amount
        - nextPayment
        - status
        - applicant
        - insured
        - periodStart
        - periodEnd
        - annualPremium
        - paymentPeriod
        - coveragePeriod
        - responsibilities
        - paymentHistory
        - policyNo